tests/wpt/mozilla/tests/mozilla/range_deleteContents.html

<!DOCTYPE html>
<html>
    <head>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
    </head>
    <body>
        <script>
            test(function() {
                var range = document.createRange();
                range.setStart(document, 0);
                range.setEnd(document, 1);
                range.deleteContents();
                assert_equals(document.childNodes.length, 1);
            }, "Deleting range containing doctype");
        </script>
    </body>
</html>
